+++ AetherBags/System.cs
@@ -13,6 +13,7 @@ public static class System
     public static AddonSaddleBagWindow AddonSaddleBagWindow { get; set; } = null!;
     public static AddonRetainerWindow AddonRetainerWindow { get; set; } = null!;
     public static AddonConfigurationWindow AddonConfigurationWindow { get; set; } = null!;
+    public static AddonChangelogWindow AddonChangelogWindow { get; set; } = null!;
     public static IPCService IPC { get; set; } = null!;
     public static AetherBagsIPCProvider? AetherBagsAPI { get; set; }
     public static SystemConfiguration Config { get; set; } = null!;